+++ build.rs
@@ -2,6 +2,11 @@ use std::env;
 
 // mostly copied from the current_platform crate.
 fn main() {
+    // We need this to build `ubi` properly in the integration tests.
+    for (key, value) in env::vars().filter(|(k, _)| k.starts_with("CARGO_FEATURE_")) {
+        println!("cargo:rustc-env={key}={value}");
+    }
+
     // Cargo sets the host and target env vars for build scripts, but not
     // crates:
     // https://doc.rust-lang.org/cargo/reference/environment-variables.html#environment-variables-cargo-sets-for-build-scripts