+++ tests/ubi.rs
@@ -54,7 +54,17 @@ impl PreservableTempdir {
 #[allow(clippy::too_many_lines)]
 fn integration_tests() -> Result<()> {
     let cargo = make_exe_pathbuf(&["cargo"]);
-    run_command(cargo.as_ref(), &["build"])?;
+
+    let features = if env::var("CARGO_FEATURE_RUSTLS_TLS_NATIVE_ROOTS").unwrap_or_default() == "1" {
+        "rustls-tls-native-roots"
+    } else if env::var("CARGO_FEATURE_NATIVE_TLS").unwrap_or_default() == "1" {
+        "native-tls"
+    } else if env::var("CARGO_FEATURE_NATIVE_TLS_VENDORED").unwrap_or_default() == "1" {
+        "native-tls-vendored"
+    } else {
+        "default"
+    };
+    run_command(&cargo, &["build", "--features", features])?;
 
     let mut ubi = env::current_dir()?;
     ubi.push("target");